+++ platform/linux/native.h
@@ -35,4 +35,6 @@ void vitra_tray_clear_menu(void);
 void vitra_tray_add_menu_item(const char *item_id, const char *item_label);
 void vitra_tray_clear(void);
 
+void vitra_win_set_drag_drop(VitraWin *w, int enabled);
+
 #endif